Added Documentation and Cleaned up Comment View Controller along with other changes.
This commit has ckeaned up CommentVC making it easier to navigate this section.
