+++ Slide for Reddit/Comments/Views/CommentCellView.swift
