+++ Slide for Reddit/PagingCommentViewController.swift
@@ -41,7 +41,7 @@ class PagingCommentViewController: ColorMuxPagingViewController, UIPageViewContr
                                    direction: .forward,
                                    animated: true,
                                    completion: nil)
-            vc.refresh(vc)
+            vc.refreshComments(vc)
         }
     }
     
@@ -133,7 +133,7 @@ class PagingCommentViewController: ColorMuxPagingViewController, UIPageViewContr
         if !(self.viewControllers!.first! is ClearVC) {
             PagingCommentViewController.savedComment = self.viewControllers!.first as? CommentViewController
             if !PagingCommentViewController.savedComment!.loaded {
-                PagingCommentViewController.savedComment!.refresh(pageViewController)
+                PagingCommentViewController.savedComment!.refreshComments(pageViewController)
             }
 
         }
@@ -201,7 +201,7 @@ class PagingCommentViewController: ColorMuxPagingViewController, UIPageViewContr
         }
 
         if !(viewController as! CommentViewController).loaded {
-            (viewController as! CommentViewController).refresh(viewController)
+            (viewController as! CommentViewController).refreshComments(viewController)
         }
         
         var nextIndex = viewControllerIndex + 1
